refactor : FSD 계층 준수하도록 useAuth를 shared로 이동
