+++ src/app/providers/index.ts
@@ -1,3 +1,3 @@
 export { QueryProvider } from './QueryProvider';
-export { AuthProvider, useAuth } from './AuthProvider';
+export { AuthProvider} from './AuthProvider';
 export { AppRouter } from './AppRouter';
