--- src/app/providers/useAuth.ts
@@ -1,10 +1,0 @@
-import { useContext } from 'react';
-import { AuthContext } from './AuthProvider';
-
-export const useAuth = () => {
-    const context = useContext(AuthContext);
-    if (context === undefined) {
-        throw new Error('useAuth must be used within an AuthProvider');
-    }
-    return context;
-};
