+++ src/features/auth/hooks/useAuthMutations.ts
@@ -2,7 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from '@tanstack/react-router';
 import { toast } from 'react-toastify';
 import axios from 'axios';
-import { useAuth } from '@/app/providers/useAuth';
+import { useAuth } from '@/shared/lib/auth';
 import type { LoginRequest, SignupRequest } from '@/entities/auth';
 
 const getErrorMessage = (error: unknown, defaultMessage: string): string => {
