+++ src/shared/lib/auth/AuthContext.ts
@@ -1,0 +1,13 @@
+import { createContext } from 'react';
+import type { User, LoginRequest, SignupRequest } from '@/entities/auth';
+
+export interface AuthContextType {
+  user: User | null;
+  isLoggedIn: boolean;
+  isLoading: boolean;
+  login: (credentials: LoginRequest) => Promise<void>;
+  signup: (userData: SignupRequest) => Promise<void>;
+  logout: () => Promise<void>;
+}
+
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
