+++ src/shared/lib/auth/index.ts
@@ -1,0 +1,2 @@
+export { AuthContext, type AuthContextType } from './AuthContext';
+export { useAuth } from './useAuth';
